fourth

+++ starter_code/backend/src/database/models.py
@@ -120,4 +120,4 @@ class Drink(db.Model):
         db.session.commit()
 
     def __repr__(self):
-        return json.dumps(self.long())
+        return json.dumps(self.short())
